main: extract object_storage_endpoint_param out
so we can reuse it in scylla-nodetool to read object-store.yaml
when uploading/downloading sstables.

Refs scylladb/scylladb#20538
